[Editorial] Remove or resemanticate italics in quotes
The two instances of direct dialogue in The Ballad of Reading Gaol are marked with both quotation marks and italics. In accordance with SE editorial policy, I have removed these from the second, longer quotation, but as an editorial judgment I recoded the first italicised phrase ("That fellow's got to swing") as emphasis because of its rhetotical impact in the poem.